requirements_fastapi.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.1
pydantic-settings==2.1.0
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.0
torch==2.7.1
torchvision==0.22.1
numpy==1.26.2
opencv-python==4.8.1.78
Pillow==10.1.0
facenet-pytorch==2.5.3
aiofiles==23.2.1
# Video processing dependencies
moviepy==2.2.1
imageio-ffmpeg==0.6.0
# Additional dependencies for multi-model system
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
tqdm==4.66.1
tensorboard==2.15.1
albumentations==1.3.1
psutil==5.9.6
email-validator==2.1.0

# Hugging Face deepfake detection model
transformers>=4.55.0
